bot/requirements.txt
aiogram==3.15.0
python-dotenv==1.0.1
httpx>=0.27.0